refactor : Command에서 뷰에 대한 로직을 제거하고 ConsoleController에서 처리하도록 변경
