+++ src/main/java/chess/ChessConsoleApplication.java
@@ -1,0 +1,10 @@
+package chess;
+
+import chess.controller.ChessConsoleController;
+
+public class ChessConsoleApplication {
+
+    public static void main(final String[] args) {
+        new ChessConsoleController().run();
+    }
+}
