--- src/main/java/chess/ConsoleApplication.java
@@ -1,8 +1,0 @@
-package chess;
-
-public class ConsoleApplication {
-
-    public static void main(final String[] args) {
-        new ChessGameLauncher().run();
-    }
-}
