+++ src/test/java/chess/domain/CommandTest.java
@@ -1,8 +1,9 @@
-package chess;
+package chess.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.domain.Command;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,31 +12,31 @@ public class CommandTest {
     @Test
     @DisplayName("start Command가 존재하는지 확인한다.")
     void start() {
-        assertThat(Command.of("start")).isEqualTo(Command.START);
+        assertThat(Command.from("start")).isEqualTo(Command.START);
     }
 
     @Test
     @DisplayName("end Command가 존재하는지 확인한다.")
     void end() {
-        assertThat(Command.of("end")).isEqualTo(Command.END);
+        assertThat(Command.from("end")).isEqualTo(Command.END);
     }
 
     @Test
     @DisplayName("move Command가 존재하는지 확인한다.")
     void move() {
-        assertThat(Command.of("move")).isEqualTo(Command.MOVE);
+        assertThat(Command.from("move")).isEqualTo(Command.MOVE);
     }
 
     @Test
     @DisplayName("status Command가 존재하는지 확인한다.")
     void status() {
-        assertThat(Command.of("status")).isEqualTo(Command.STATUS);
+        assertThat(Command.from("status")).isEqualTo(Command.STATUS);
     }
 
     @Test
     @DisplayName("존재하지 않는 Command를 입력했을경우 예외를 발생시킨다.")
     void checkNonCommandException() {
-        assertThatThrownBy(() -> Command.of("go"))
+        assertThatThrownBy(() -> Command.from("go"))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("[ERROR] 존재하지 않는 명령어입니다.");
     }
